♻️ refactor(cn): changing the import location
