+++ frontend/components/AbbreviatedAvatar.tsx
@@ -5,7 +5,7 @@ import type { AvatarProps } from '@nextui-org/react'
 import React from 'react'
 import { Avatar } from '@nextui-org/react'
 
-import { cn } from '@/utils/nextui/cn'
+import { cn } from '@/utils/cn'
 
 const AbbreviatedAvatar = React.forwardRef<HTMLSpanElement, AvatarProps>(
   ({ name, className, classNames = {}, ...props }, ref) => (
