+++ frontend/features/alert/components/HistoryAlert.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { cn } from '@/utils/nextui/cn'
+import { cn } from '@/utils/cn'
 import { Icon } from '@iconify/react'
 import {
   Button,
